gemini/OfAKind-Python/requirements.txt
streamlit
python-dotenv
google-generativeai
streamlit-geolocation
